dedup: share IL operand token resolution (fix IL vs Annotated-IL divergence)
ILDisassembler and AnnotatedILEmitter each had their own copy of IL metadata-token
resolution (~150 lines), and they had drifted: ILDisassembler rendered call operands
with full parameter lists and generic type args (System.Version::ToString(int)) while
AnnotatedILEmitter emitted empty parens (System.Version::ToString()), so the IL and
"IL (Annotated)" sections of the same `api`/`member` output disagreed on the same call.

Extract the canonical resolution into Metadata.ILTokenResolver
(ResolveString/Type/Method/Field/Token + Format* + EscapeString). Type/token
resolution takes an optional GenericContext so the annotated emitter keeps resolving
!0/!!0 from the method body context, while the standalone disassembler passes null and
its output stays byte-identical. AnnotatedILEmitter now inherits the full-parameter
formatting, so both sections agree.

Verified on System.Version.ToString(): both sections now show
`call System.Version::ToString(int)`. All suites green (Decompiler 208, Metadata 131,
CLI 919).
