+++ src/DotnetInspector.Metadata/ILDisassembler.cs
@@ -127,11 +127,11 @@ public static class ILDisassembler
             OperandKind.R => ReadDouble(ilBytes, ref position).ToString(),
             OperandKind.ShortVariable => ReadByte(ilBytes, ref position).ToString(),
             OperandKind.Variable => ReadUInt16(ilBytes, ref position).ToString(),
-            OperandKind.String => ResolveString(ReadInt32(ilBytes, ref position), reader),
-            OperandKind.Type => ResolveType(ReadInt32(ilBytes, ref position), reader),
-            OperandKind.Method => ResolveMethod(ReadInt32(ilBytes, ref position), reader),
-            OperandKind.Field => ResolveField(ReadInt32(ilBytes, ref position), reader),
-            OperandKind.Tok => ResolveToken(ReadInt32(ilBytes, ref position), reader),
+            OperandKind.String => ILTokenResolver.ResolveString(reader, ReadInt32(ilBytes, ref position)),
+            OperandKind.Type => ILTokenResolver.ResolveType(reader, ReadInt32(ilBytes, ref position)),
+            OperandKind.Method => ILTokenResolver.ResolveMethod(reader, ReadInt32(ilBytes, ref position)),
+            OperandKind.Field => ILTokenResolver.ResolveField(reader, ReadInt32(ilBytes, ref position)),
+            OperandKind.Tok => ILTokenResolver.ResolveToken(reader, ReadInt32(ilBytes, ref position)),
             OperandKind.Sig => $"0x{ReadInt32(ilBytes, ref position):X8}",
             OperandKind.Switch => DecodeSwitch(ilBytes, ref position),
             _ => null
@@ -155,232 +155,6 @@ public static class ILDisassembler
             targets.Add($"IL_{(baseOffset + delta):X4}");
         }
         return $"({string.Join(", ", targets)})";
-    }
-
-    // Token resolution
-
-    static string ResolveString(int token, MetadataReader reader)
-    {
-        try
-        {
-            var handle = MetadataTokens.UserStringHandle(token);
-            var value = reader.GetUserString(handle);
-            return $"\"{EscapeString(value)}\"";
-        }
-        catch
-        {
-            return $"0x{token:X8}";
-        }
-    }
-
-    static string ResolveType(int token, MetadataReader reader)
-    {
-        try
-        {
-            var handle = MetadataTokens.EntityHandle(token);
-            return TypeResolver.GetTypeName(reader, handle) ?? $"0x{token:X8}";
-        }
-        catch
-        {
-            return $"0x{token:X8}";
-        }
-    }
-
-    static string ResolveMethod(int token, MetadataReader reader)
-    {
-        try
-        {
-            var handle = MetadataTokens.EntityHandle(token);
-            return handle.Kind switch
-            {
-                HandleKind.MethodDefinition => FormatMethodDef(reader, (MethodDefinitionHandle)handle),
-                HandleKind.MemberReference => FormatMemberRef(reader, (MemberReferenceHandle)handle),
-                HandleKind.MethodSpecification => FormatMethodSpec(reader, (MethodSpecificationHandle)handle),
-                _ => $"0x{token:X8}"
-            };
-        }
-        catch
-        {
-            return $"0x{token:X8}";
-        }
-    }
-
-    static string ResolveField(int token, MetadataReader reader)
-    {
-        try
-        {
-            var handle = MetadataTokens.EntityHandle(token);
-            return handle.Kind switch
-            {
-                HandleKind.FieldDefinition => FormatFieldDef(reader, (FieldDefinitionHandle)handle),
-                HandleKind.MemberReference => FormatMemberRef(reader, (MemberReferenceHandle)handle),
-                _ => $"0x{token:X8}"
-            };
-        }
-        catch
-        {
-            return $"0x{token:X8}";
-        }
-    }
-
-    static string ResolveToken(int token, MetadataReader reader)
-    {
-        try
-        {
-            var handle = MetadataTokens.EntityHandle(token);
-            return handle.Kind switch
-            {
-                HandleKind.TypeDefinition or HandleKind.TypeReference or HandleKind.TypeSpecification
-                    => TypeResolver.GetTypeName(reader, handle) ?? $"0x{token:X8}",
-                HandleKind.MethodDefinition => FormatMethodDef(reader, (MethodDefinitionHandle)handle),
-                HandleKind.MemberReference => FormatMemberRef(reader, (MemberReferenceHandle)handle),
-                HandleKind.FieldDefinition => FormatFieldDef(reader, (FieldDefinitionHandle)handle),
-                _ => $"0x{token:X8}"
-            };
-        }
-        catch
-        {
-            return $"0x{token:X8}";
-        }
-    }
-
-    // Formatting helpers
-
-    static string FormatMethodDef(MetadataReader reader, MethodDefinitionHandle handle)
-    {
-        var method = reader.GetMethodDefinition(handle);
-        string methodName = reader.GetString(method.Name);
-        var declaringType = method.GetDeclaringType();
-        string typeName = TypeResolver.GetTypeName(reader, declaringType) ?? "?";
-        string baseName = $"{typeName}::{methodName}";
-
-        try
-        {
-            var sig = method.DecodeSignature(SignatureDecoder.Instance, genericContext: null);
-            return $"{baseName}({string.Join(", ", sig.ParameterTypes)})";
-        }
-        catch
-        {
-            return baseName;
-        }
-    }
-
-    static string FormatMemberRef(MetadataReader reader, MemberReferenceHandle handle)
-    {
-        var memberRef = reader.GetMemberReference(handle);
-        string name = reader.GetString(memberRef.Name);
-        string? parentName = TypeResolver.GetTypeName(reader, memberRef.Parent);
-        string baseName = parentName is not null ? $"{parentName}::{name}" : name;
-
-        try
-        {
-            if (memberRef.GetKind() == MemberReferenceKind.Method)
-            {
-                var genericCtx = BuildGenericContext(reader, memberRef);
-                var sig = memberRef.DecodeMethodSignature(SignatureDecoder.Instance, genericCtx);
-                return $"{baseName}({string.Join(", ", sig.ParameterTypes)})";
-            }
-        }
-        catch
-        {
-            // Fall through to baseName
-        }
-
-        return baseName;
-    }
-
-    static string FormatMethodSpec(MetadataReader reader, MethodSpecificationHandle handle)
-    {
-        var spec = reader.GetMethodSpecification(handle);
-        string baseName = spec.Method.Kind switch
-        {
-            HandleKind.MethodDefinition => FormatMethodDef(reader, (MethodDefinitionHandle)spec.Method),
-            HandleKind.MemberReference => FormatMemberRef(reader, (MemberReferenceHandle)spec.Method),
-            _ => $"0x{MetadataTokens.GetToken(spec.Method):X8}"
-        };
-
-        try
-        {
-            var typeArgs = spec.DecodeSignature(SignatureDecoder.Instance, genericContext: null);
-            return $"{baseName}<{string.Join(", ", typeArgs)}>";
-        }
-        catch
-        {
-            return baseName;
-        }
-    }
-
-    static string FormatFieldDef(MetadataReader reader, FieldDefinitionHandle handle)
-    {
-        var field = reader.GetFieldDefinition(handle);
-        string fieldName = reader.GetString(field.Name);
-        var declaringType = field.GetDeclaringType();
-        string typeName = TypeResolver.GetTypeName(reader, declaringType) ?? "?";
-        return $"{typeName}::{fieldName}";
-    }
-
-    static string EscapeString(string value)
-    {
-        return value
-            .Replace("\\", "\\\\")
-            .Replace("\"", "\\\"")
-            .Replace("\n", "\\n")
-            .Replace("\r", "\\r")
-            .Replace("\t", "\\t");
-    }
-
-    /// <summary>
-    /// Build a GenericContext from a MemberReference's parent type for accurate
-    /// parameter type rendering on generic type members.
-    /// </summary>
-    static GenericContext? BuildGenericContext(MetadataReader reader, MemberReference memberRef)
-    {
-        try
-        {
-            if (memberRef.Parent.Kind == HandleKind.TypeSpecification)
-            {
-                var typeSpec = reader.GetTypeSpecification((TypeSpecificationHandle)memberRef.Parent);
-                var parentType = typeSpec.DecodeSignature(SignatureDecoder.Instance, genericContext: null);
-                var typeArgs = ExtractGenericArguments(parentType);
-                if (typeArgs.Count > 0)
-                    return new GenericContext(typeArgs, []);
-            }
-        }
-        // Graceful fallback when generic type signature cannot be decoded
-        catch { }
-        return null;
-    }
-
-    static IReadOnlyList<string> ExtractGenericArguments(string typeName)
-    {
-        int openAngle = typeName.IndexOf('<');
-        if (openAngle < 0) return [];
-
-        int depth = 0;
-        List<string> args = [];
-        int argStart = openAngle + 1;
-
-        for (int i = openAngle; i < typeName.Length; i++)
-        {
-            switch (typeName[i])
-            {
-                case '<': depth++; break;
-                case '>':
-                    depth--;
-                    if (depth == 0)
-                    {
-                        var arg = typeName[argStart..i].Trim();
-                        if (arg.Length > 0) args.Add(arg);
-                        return args;
-                    }
-                    break;
-                case ',' when depth == 1:
-                    args.Add(typeName[argStart..i].Trim());
-                    argStart = i + 1;
-                    break;
-            }
-        }
-        return args;
     }
 
     // Primitive readers
